% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion_filter_module.R
\name{accordionFilterModuleServer}
\alias{accordionFilterModuleServer}
\title{accordion filter UI module element}
\usage{
accordionFilterModuleServer(id, data, filterVars)
}
\arguments{
\item{id}{module id string}

\item{data}{dataframe to be filtered}

\item{filterVars}{columns names of}
}
\value{
a filtered version of the dataset as a reactive
}
\description{
accordion filter UI module element
}
