% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accordion_filter_module.R
\name{accordionFilterModuleUI}
\alias{accordionFilterModuleUI}
\title{accordion filter UI module element}
\usage{
accordionFilterModuleUI(id, style = NULL)
}
\arguments{
\item{id}{module id string}

\item{style}{css style properties to be applied to a div containing the widget}
}
\value{
ui elements of the widget
}
\description{
accordion filter UI module element
}
