% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterPie.R
\name{filterPie}
\alias{filterPie}
\title{pie chart filter widget}
\usage{
filterPie(
  data,
  options = list(),
  reactiveID = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{data list object to be plotted; typically generated using \code{dataFormat()}.}

\item{options}{list of apache echarts options}

\item{reactiveID}{string id to be passed to \code{Shiny.setInputValue()} called within filterPie.js}

\item{width}{width of the graph as CSS units.}

\item{height}{height of the graph as CSS units.}

\item{elementId}{html element id.}
}
\description{
an apache pie chart that can be clicked to apply filters within an R shiny app.
}
\examples{

rawVector <- sample(c("French", "German", "British", "American", "Canadian", "Dutch"),
1000, replace=TRUE, prob=c(0.2, 0.3, 0.3, .1, .05, .05))

rawVector |>
   dataFormatPie() |>
   filterPie()

}
